test(pro-pay): auto config pass on missing vp, zl beans
